Bug 817035: Major UI refactor.
The visible changes here:

 * the toolbar is now "tabbed", albeit with a CSS implementation of tabs
   that will make your eyes bleed.  Sorry.
 * Lifeguard and BMM display different columns - we need the space!
 * Can set comments in the web UI
 * Hopefully a clearer delineation of PXE configs, bootconfig, and b2gbase

The invisible change is a significant refactoring of the views code to
dynamically generate each "action" tab, using more generic View classes
rather than purpose-specific.

All of the functionality remains the same.

Add support for setting comments in the web UI.

This also adds some help text to some of the views.