Bug 817035: Add environments and allow requests to specify one
This adds 'environment' columns to requests and devices, and allows the
environment to be viewed and edited on the BMM page.

This also *significantly* refactors updatable models and the views,
splitting the views out into several JS files (since views.js was
getting large).

NOTE: The request API is backward-compatible -- if no environment is
specified, it defaults to 'any', with the same semantics as before.